Fix clone copying attributes
