+++ agilerl/algorithms/core/base.py
@@ -1,6 +1,10 @@
 import copy
 import inspect
 import warnings
+import os 
+import glob
+import gc
+import deepspeed
 from abc import ABC, ABCMeta, abstractmethod
 from importlib.metadata import version
 from typing import (
@@ -58,6 +62,9 @@ from agilerl.utils.algo_utils import (
     preprocess_observation,
     recursive_check_module_attrs,
     remove_compile_prefix,
+    remove_nested_files,
+    clone_llm,
+    is_peft_model
 )
 from agilerl.utils.evolvable_networks import is_image_space
 
@@ -1427,3 +1434,236 @@ class MultiAgentRLAlgorithm(EvolvableAlgorithm, ABC):
             )
             summed_rewards[homo_id] += reward
         return summed_rewards
+    
+
+class LLMAlgorithm(EvolvableAlgorithm, ABC):
+    """Base object for all LLM algorithms in the AgileRL framework.
+    
+    """
+    def __init__(self, 
+                 observation_space: spaces.Space, 
+                 action_space: spaces.Space, 
+                 index: int, 
+                 hp_config: Optional[HyperparameterConfig] = None, 
+                 device: Union[str, torch.device] = "cpu", 
+                 accelerator: Optional[Accelerator] = None, 
+                 name: Optional[str] = None) -> None:
+        super().__init__(index, hp_config, device, accelerator, None, name)
+        assert isinstance(
+            observation_space, spaces.Space
+        ), "Observation space must be an instance of gymnasium.spaces.Space."
+        assert isinstance(
+            action_space, spaces.Space
+        ), "Action space must be an instance of gymnasium.spaces.Space."
+
+        self.observation_space = observation_space
+        self.action_space = action_space
+        self.zero_stage = None
+        if self.accelerator is not None:
+            self.zero_stage = self.accelerator.state.deepspeed_plugin.deepspeed_config["zero_optimization"]["stage"]
+        if self.zero_stage == 3 and self.accelerator.is_main_process:
+            warnings.warn(
+                "Zero stage 3 is feature is nascent and has not been thoroughly tested. It may be unstable, incomplete, or subject to change. Use at your own risk. We recommend caution in production environments."
+            )
+
+
+    def preprocess_observation(self, observation: ObservationType) -> TorchObsType:
+        """Dummy preprocesses observations for forward pass through neural network.
+
+        :param observations: Observations of environment
+        :type observations: numpy.ndarray[float] or dict[str, numpy.ndarray[float]]
+
+        :return: Preprocessed observations
+        :rtype: torch.Tensor[float] or dict[str, torch.Tensor[float]] or Tuple[torch.Tensor[float], ...]
+        """
+        return observation
+    
+    def save_checkpoint(self, path: str) -> None:
+        """
+        Override the save_checkpoint method to provide guidance on the correct method to use.
+        :param path: Location to save checkpoint at
+        :type path: string
+        """
+        raise NotImplementedError(
+            "The save_checkpoint method is not supported for this algorithm class. "
+            "Please use agent.actor.save_pretrained(checkpoint_path) instead."
+        )
+
+    def load_checkpoint(self, path: str) -> None:
+        """
+        Override the load_checkpoint method to provide guidance on the correct method to use.
+
+        :param path: Location to load checkpoint from
+        :type path: string
+        """
+        raise NotImplementedError(
+            "The load_checkpoint method is not supported for this algorithm class."
+            """
+            To load a saved LLM, please load the model as follows, and then re-instantiate the GRPO
+            class.
+
+            base_model = AutoModelForCausalLM.from_pretrained(
+                "Qwen/Qwen2.5-3B",
+                torch_dtype=torch.bfloat16,
+                device_map="auto"
+            )
+            tokenizer = AutoTokenizer.from_pretrained("Qwen/Qwen2.5-3B")
+            model = PeftModel.from_pretrained(base_model, "/path/to/adapter/folder")
+            """
+        )
+
+    def _save_distributed_actor(self, path: str) -> None:
+        """
+        Override the save_checkpoint method to provide guidance on the correct method to use.
+
+        :param path: Output directory to save the checkpoint at
+        :type path: str
+        """
+        if self.accelerator is not None:
+            os.makedirs(path, exist_ok=True)
+            self.actor.save_checkpoint(path, tag="checkpoint")    
+        else:
+            warnings.warn(
+                "Distributed actor save not supported for non-distributed training."
+            )
+
+    def _load_distributed_actor(self, path: str) -> None:
+        """
+        Override the load_checkpoint method to provide guidance on the correct method to use.
+
+        :param path: Output directory to load the checkpoint from
+        :type path: str
+        """
+        if self.accelerator is not None:
+            deepspeed_dirs = sorted(glob.glob(f"{path}/checkpoint"))
+            assert len(deepspeed_dirs) > 0
+            load_path, _ = self.actor.load_checkpoint(
+                path,
+                tag="checkpoint",
+                load_module_strict=not is_peft_model(self.accelerator.unwrap_model(self.actor)), 
+                load_optimizer_states=True,
+                load_lr_scheduler_states=True,
+            )
+
+            if load_path is None:
+                raise ValueError(f"[deepspeed] failed to resume from checkpoint {path}")
+        else:
+            warnings.warn(
+                "Distributed actor load not supported for non-distributed training."
+            )
+
+    @classmethod
+    def load(
+        cls,
+        path: str,
+        device: DeviceType = "cpu",
+        accelerator: Optional[Accelerator] = None,
+    ) -> None:
+        raise NotImplementedError(
+            "The load class method is not supported for this algorithm class."
+            """
+            To load a saved LLM, please load the model as follows, and then re-instantiate the GRPO
+            class, using the pre-trained model.
+
+            base_model = AutoModelForCausalLM.from_pretrained(
+                "Qwen/Qwen2.5-3B",
+                torch_dtype=torch.bfloat16,
+                device_map="auto"
+            )
+            tokenizer = AutoTokenizer.from_pretrained("Qwen/Qwen2.5-3B")
+            model = PeftModel.from_pretrained(base_model, "/path/to/adapter/folder")
+            """
+        )
+
+    def wrap_models(self):
+        """Wrap the models in the accelerator
+        """
+        if self.accelerator is not None:
+            self.actor, self.optimizer, self.lr_scheduler = self.accelerator.prepare(
+                self.actor, self.optimizer.optimizer, self.lr_scheduler
+            )
+            deepspeed_plugin = self.accelerator.state.deepspeed_plugin
+            config_kwargs = copy.deepcopy(deepspeed_plugin.deepspeed_config)
+            config_kwargs["zero_optimization"]["stage"] = 0
+            self.reference_actor, *_ = deepspeed.initialize(
+                model=self.reference_actor, config=config_kwargs
+            )
+
+
+    def clone(self, index: Optional[int] = None, wrap: bool = True):
+        """Creates a clone of the algorithm.
+
+        :param index: The index of the clone, defaults to None
+        :type index: Optional[int], optional
+        :param wrap: If True, wrap the models in the clone with the accelerator, defaults to False
+        :type wrap: bool, optional
+
+        :return: A clone of the algorithm
+        :rtype: EvolvableAlgorithm
+        """
+
+        if self.zero_stage == 3:
+            self.accelerator.wait_for_everyone()
+            self._save_distributed_actor(f"temporary_checkpoint/agent_{self.index}")
+            self.accelerator.wait_for_everyone()
+
+        input_args = EvolvableAlgorithm.inspect_attributes(
+                self, input_args_only=True
+        )
+        input_args["clone"] = True
+
+        # extract base model and peft config
+        actor = self.accelerator.unwrap_model(self.actor) if self.accelerator is not None else self.actor
+        cloned_actor = clone_llm(actor, load_state_dict=(self.zero_stage != 3)) # NOTE do we want to load this state dict given we load the checkpoint in?
+
+        input_args["actor_network"] = cloned_actor
+        input_args["accelerator"] = Accelerator()
+        clone = type(self)(**input_args)
+
+        clone.reference_actor.load_state_dict(self.reference_actor.state_dict())
+        clone.reference_actor.eval()
+
+        clone.mutation_hook()
+
+        # Clone attributes
+        accelerator = clone.accelerator 
+        clone = EvolvableAlgorithm.copy_attributes(self, clone)
+        clone.accelerator = accelerator
+
+        if self.accelerator is None:    
+            clone.optimizer.optimizer.load_state_dict(self.optimizer.optimizer.state_dict())
+            if self.lr_scheduler is not None:
+                clone.lr_scheduler.load_state_dict(self.lr_scheduler.state_dict())
+        # Set the index
+        if index is not None:
+            clone.index = index
+
+        if self.zero_stage == 3:
+            clone.accelerator.wait_for_everyone()
+            clone._load_distributed_actor(f"temporary_checkpoint/agent_{self.index}")
+            clone.accelerator.wait_for_everyone()
+            saved_state_files = glob.glob(f"temporary_checkpoint/agent_{self.index}/*")
+            clone.accelerator.wait_for_everyone()
+            if self.accelerator.is_main_process:
+                remove_nested_files(saved_state_files)
+
+        return clone
+
+
+    def __del__(self) -> None:
+        """Delete the algorithm.
+
+        :return: None
+        :rtype: None
+        """
+        if self.accelerator is not None:
+            self.accelerator.free_memory()
+            self.accelerator.wait_for_everyone()
+        del self.reference_actor 
+        del self.actor 
+        del self.optimizer
+        del self.lr_scheduler
+        gc.collect()
+        torch.cuda.empty_cache()
+
+
