+++ benchmarking/benchmarking_grpo.py
@@ -34,18 +34,19 @@ def create_model(pretrained_model_name_or_path):
         pretrained_model_name_or_path=pretrained_model_name_or_path,
         torch_dtype=torch.bfloat16,
         attn_implementation="flash_attention_2",
+        # quantization_config=bnb_config,
     )
     peft_config = LoraConfig(
-        r=32,
-        lora_alpha=32,
+        r=16,
+        lora_alpha=64,
         target_modules=[
             "q_proj",
             "k_proj",
             "v_proj",
-            "o_proj",
-            "up_proj",
-            "down_proj",
-            "gate_proj",
+            # "o_proj",
+            # "up_proj",
+            # "down_proj",
+            # "gate_proj",
         ],
         task_type="CAUSAL_LM",
         lora_dropout=0.05,
